move: Animation 파일 rules 패키지로 이동
