+++ compose/src/main/java/com/yourssu/design/system/compose/rule/Animation.kt
@@ -1,4 +1,4 @@
-package com.yourssu.design.system.compose.foundation
+package com.yourssu.design.system.compose.rule
 
 import androidx.compose.animation.core.CubicBezierEasing
